+++ CHANGELOG.md
@@ -6,6 +6,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## [UNRELEASED]
+### Added
+ - Implemented cash rounding for currency formatting.
+ - Currency options now accepts `cash: boolean` to enable cash rounding.
+
 ## [0.13.6]
 ### Fixed/Changed
  - Drop utf-8 encoding before checksumming the schema config.